Extract common fields from CreateBreweryStyle and UpdateBreweryStyle commands and validators to abstract commands and validators
