+++ src/Application/Breweries/Commands/Common/BaseBreweryCommand.cs
@@ -1,0 +1,57 @@
+﻿namespace Application.Breweries.Commands.Common;
+
+/// <summary>
+///     BaseBreweryCommand abstract record.
+/// </summary>
+public abstract record BaseBreweryCommand
+{
+    /// <summary>
+    ///     The brewery name.
+    /// </summary>
+    public string? Name { get; init; }
+
+    /// <summary>
+    ///     The description.
+    /// </summary>
+    public string? Description { get; init; }
+
+    /// <summary>
+    ///     The foundation year.
+    /// </summary>
+    public int FoundationYear { get; init; }
+
+    /// <summary>
+    ///     The website url.
+    /// </summary>
+    public string? WebsiteUrl { get; init; }
+
+    /// <summary>
+    ///     The street.
+    /// </summary>
+    public string? Street { get; init; }
+
+    /// <summary>
+    ///     The house number.
+    /// </summary>
+    public string? Number { get; init; }
+
+    /// <summary>
+    ///     The post code.
+    /// </summary>
+    public string? PostCode { get; init; }
+
+    /// <summary>
+    ///     The city.
+    /// </summary>
+    public string? City { get; init; }
+
+    /// <summary>
+    ///     The state.
+    /// </summary>
+    public string? State { get; init; }
+
+    /// <summary>
+    ///     The country.
+    /// </summary>
+    public string? Country { get; init; }
+}
