+++ src/Application/Breweries/Commands/UpdateBrewery/UpdateBreweryCommandValidator.cs
@@ -1,4 +1,4 @@
-﻿using System.Text.RegularExpressions;
+﻿using Application.Breweries.Commands.Common;
 using Application.Common.Interfaces;
 using FluentValidation;
 using Microsoft.EntityFrameworkCore;
@@ -8,7 +8,7 @@ namespace Application.Breweries.Commands.UpdateBrewery;
 /// <summary>
 ///     UpdateBreweryCommand validator.
 /// </summary>
-public class UpdateBreweryCommandValidator : AbstractValidator<UpdateBreweryCommand>
+public class UpdateBreweryCommandValidator : BaseBreweryCommandValidator<UpdateBreweryCommand>
 {
     /// <summary>
     ///     The database context.
@@ -20,24 +20,11 @@ public class UpdateBreweryCommandValidator : AbstractValidator<UpdateBreweryComm
     /// </summary>
     /// <param name="context">The database context</param>
     /// <param name="dateTime">The date time service</param>
-    public UpdateBreweryCommandValidator(IApplicationDbContext context, IDateTime dateTime)
+    public UpdateBreweryCommandValidator(IApplicationDbContext context, IDateTime dateTime) : base(dateTime)
     {
         _context = context;
 
-        RuleFor(x => x.Description).NotEmpty().MaximumLength(5000);
-        RuleFor(x => x.FoundationYear).NotEmpty().InclusiveBetween(0, dateTime.Now.Year);
-        RuleFor(x => x.WebsiteUrl).MaximumLength(200).Must(url => Uri.TryCreate(url, UriKind.Absolute, out _))
-            .WithMessage("Invalid URL.");
-        RuleFor(x => x.Street).NotEmpty().MaximumLength(200);
-        RuleFor(x => x.Number).NotEmpty().MaximumLength(10);
-        RuleFor(x => x.PostCode).NotEmpty().Must(BeAValidPostalCode)
-            .WithMessage("Invalid postal code.");
-        RuleFor(x => x.City).NotEmpty().MaximumLength(50);
-        RuleFor(x => x.State).NotEmpty().MaximumLength(50);
-        RuleFor(x => x.Country).NotEmpty().MaximumLength(50);
-        RuleFor(x => x.Name).NotEmpty().MaximumLength(500)
-            .MustAsync(BeUniquelyNamed)
-            .WithMessage("The brewery name must be unique.");
+        RuleFor(x => x.Name).MustAsync(BeUniquelyNamed!).WithMessage(UniqueNameErrorMessage);
     }
 
     /// <summary>
@@ -51,21 +38,5 @@ public class UpdateBreweryCommandValidator : AbstractValidator<UpdateBreweryComm
     {
         return await _context.Breweries.Where(x => x.Id != model.Id)
             .AllAsync(x => x.Name != name.Trim(), cancellationToken);
-    }
-
-    /// <summary>
-    ///     The custom rule indicating whether postcode is valid.
-    /// </summary>
-    /// <param name="postCode">The post code</param>
-    private static bool BeAValidPostalCode(string? postCode)
-    {
-        // USA format (5 digits followed by an optional dash and 4 more digits)
-        const string usaPattern = @"^\d{5}(?:[-\s]\d{4})?$";
-
-        // Poland format (2 digits, dash, 3 digits)
-        const string polandPattern = @"^\d{2}-\d{3}$";
-
-        return !string.IsNullOrEmpty(postCode) &&
-               (Regex.IsMatch(postCode, usaPattern) || Regex.IsMatch(postCode, polandPattern));
     }
 }
